+++ board-service/src/main/java/org/example/boardservice/utils/clock/ClockHolderImpl.java
@@ -1,4 +1,4 @@
-package org.example.boardservice.utils;
+package org.example.boardservice.utils.clock;
 
 import org.springframework.stereotype.Component;
