+++ board-service/src/main/java/org/example/boardservice/utils/uuid/UuidHolder.java
@@ -1,4 +1,4 @@
-package org.example.boardservice.utils;
+package org.example.boardservice.utils.uuid;
 
 public interface UuidHolder {
     String random();
