+++ board-service/src/main/java/org/example/boardservice/utils/uuid/UuidHolderImpl.java
@@ -1,4 +1,4 @@
-package org.example.boardservice.utils;
+package org.example.boardservice.utils.uuid;
 
 import org.springframework.stereotype.Component;
