+++ board-service/src/test/java/org/example/boardservice/lab/controller/LabControllerTest.java
@@ -2,17 +2,15 @@ package org.example.boardservice.lab.controller;
 
 import org.example.boardservice.board.domain.Lab;
 import org.example.boardservice.board.domain.Star;
-import org.example.boardservice.board.dto.request.LabAddStarRequest;
-import org.example.boardservice.board.dto.request.LabRemoveStarRequest;
-import org.example.boardservice.board.dto.request.LabUpdateStarRequest;
+import org.example.boardservice.board.dto.request.lab.LabAddStarRequest;
+import org.example.boardservice.board.dto.request.lab.LabRemoveStarRequest;
+import org.example.boardservice.board.dto.request.lab.LabUpdateStarRequest;
 import org.example.boardservice.board.dto.response.RestResponse;
-import org.example.boardservice.board.infrastructure.repository.lab.LabRepository;
-import org.example.boardservice.lab.domain.LabDomainTest;
 import org.example.boardservice.mock.TestClockHolder;
 import org.example.boardservice.mock.TestContainer;
 import org.example.boardservice.mock.TestUuidHolder;
-import org.example.boardservice.utils.ClockHolder;
-import org.example.boardservice.utils.UuidHolder;
+import org.example.boardservice.utils.clock.ClockHolder;
+import org.example.boardservice.utils.uuid.UuidHolder;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.springframework.http.HttpStatus;
