+++ board-service/src/test/java/org/example/boardservice/mock/TestClockHolder.java
@@ -2,7 +2,7 @@ package org.example.boardservice.mock;
 
 //import lombok.RequiredArgsConstructor;
 import lombok.RequiredArgsConstructor;
-import org.example.boardservice.utils.ClockHolder;
+import org.example.boardservice.utils.clock.ClockHolder;
 
 @RequiredArgsConstructor
 public class TestClockHolder implements ClockHolder {
