+++ board-service/src/test/java/org/example/boardservice/mock/TestUuidHolder.java
@@ -1,7 +1,7 @@
 package org.example.boardservice.mock;
 
 import lombok.RequiredArgsConstructor;
-import org.example.boardservice.utils.UuidHolder;
+import org.example.boardservice.utils.uuid.UuidHolder;
 
 @RequiredArgsConstructor
 public class TestUuidHolder implements UuidHolder {
